wip day04

+++ day_02/ex02/srcs/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: agiraude <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/18 16:57:44 by agiraude          #+#    #+#             */
-/*   Updated: 2022/09/20 12:54:23 by agiraude         ###   ########.fr       */
+/*   Updated: 2022/09/21 08:58:26 by agiraude         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -96,7 +96,7 @@ Fixed	Fixed::operator*(Fixed const & rhs) const
 {
 	Fixed	x;
 
-	x.setRawBits((long(this->_n) * long(rhs.getRawBits())) >> this->_fNbBits);
+	x.setRawBits((long(this->_n) * long(rhs.getRawBits())) >> long(this->_fNbBits));
 	return x;
 }
 
@@ -104,7 +104,7 @@ Fixed	Fixed::operator/(Fixed const & rhs) const
 {
 	Fixed	x;
 
-	x.setRawBits((long(this->_n) << this->_fNbBits) / rhs.getRawBits());
+	x.setRawBits((long(this->_n) << long(this->_fNbBits)) / rhs.getRawBits());
 	return x;
 }
